requirements.txt
Django
django_extensions
django_registration